+++ nuxeo-drive-client/nxdrive/engine/next/engine_next.py
@@ -2,25 +2,14 @@
 """ Evolution to try new engine solution. """
 
 from logging import getLogger
-from nxdrive.client.remote_document_client import RemoteDocumentClient
-from nxdrive.client.remote_file_system_client import RemoteFileSystemClient
-from nxdrive.client.remote_filtered_file_system_client import \
-    RemoteFilteredFileSystemClient
+
 from nxdrive.engine.engine import Engine
-from nxdrive.options import Options
 
 log = getLogger(__name__)
 
 
 class EngineNext(Engine):
-
-    def __init__(self, manager, definition, binder=None, processors=5,
-                 remote_doc_client_factory=RemoteDocumentClient,
-                 remote_fs_client_factory=RemoteFileSystemClient,
-                 remote_filtered_fs_client_factory=RemoteFilteredFileSystemClient):
-        super(EngineNext, self).__init__(manager, definition, binder, processors,
-                 remote_doc_client_factory, remote_fs_client_factory, remote_filtered_fs_client_factory)
-        self._type = "NXDRIVENEXT"
+    type = 'NXDRIVENEXT'
 
     def create_processor(self, item_getter, name=None):
         from nxdrive.engine.next.processor import Processor
@@ -28,6 +17,8 @@ class EngineNext(Engine):
 
     def _create_queue_manager(self, processors):
         from nxdrive.engine.next.queue_manager import QueueManager
+        from nxdrive.options import Options
+
         if Options.debug:
             return QueueManager(self, self._dao, max_file_processors=2)
         return QueueManager(self, self._dao)